requirements.txt
mcp
google-adk
monocle_apptrace>=0.4.0